pyproject.toml
[project]
name = "zyzntai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "captcha>=0.7.1",
    "coloredlogs>=15.0.1",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "langchain-text-splitters>=0.3.8",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.6",
    "pymilvus>=2.5.7",
    "pymysql>=1.1.1",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
]
